[TASK] Update for 7.6
