+++ ChangeLog
@@ -1,3 +1,11 @@
+2016-06-09 Joerg Kummer <jkummer@3pc.de>
+
+	* [TASK] Update for 7.6
+	
+2015-01-14 Alex Kellner <https://github.com/einpraegsam/ad_rtepasteplain>
+
+	* [TASK] Update for 6.2
+	
 2010-01-30 Alexander Dick  <alex@dick.at>
 
 	* initial code generated with kickstarter
