+++ ext_emconf.php
@@ -12,8 +12,7 @@
 
 $EM_CONF[$_EXTKEY] = array(
 	'title' => 'htmlArea RTE - Paste As Plain Text',
-	'description' => 'This extension forces htmlArea RTE to automatically paste the 
-clipboard content as plain text. Developed and tested with rtehtmlarea 1.8.6 and Typo3 4.3.0',
+	'description' => 'This extension forces htmlArea RTE to automatically paste the clipboard content as plain text. Developed and tested with rtehtmlarea and TYPO3 7.6.0',
 	'category' => 'be',
 	'version' => '2.1.6',
 	'state' => 'beta',
@@ -25,7 +24,7 @@ clipboard content as plain text. Developed and tested with rtehtmlarea 1.8.6 and
 	'author_company' => '',
 	'constraints' => array(
 		'depends' => array(
-			'typo3' => '6.2.0-6.2.99'
+			'typo3' => '7.6.0-7.6.99'
 		),
 		'conflicts' => array(),
 		'suggests' => array(),
