+++ ext_localconf.php
@@ -1,7 +1,8 @@
 <?php
-if (!defined ('TYPO3_MODE')) 
-	die ('Access denied.');
+if (!defined('TYPO3_MODE')) {
+    die ('Access denied.');
+}
 
-$GLOBALS['TYPO3_CONF_VARS']['SYS']['Objects']['TYPO3\\CMS\\Rtehtmlarea\\Extension\\RemoveFormat'] = array(
-	'className' => 'In2code\\AdRtepasteplain\\Extension\\RemoveFormat',
+$GLOBALS['TYPO3_CONF_VARS']['SYS']['Objects']['TYPO3\\CMS\\Rtehtmlarea\\Form\\Element\\RichTextElement'] = array(
+    'className' => 'TYPO3\\CMS\\AdRtepasteplain\\Xclass\\RichTextElement',
 );
